LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.object.update-expected.txt

2d.gradient.object.update
Actual output:
Expected output:


FAIL Canvas test: 2d.gradient.object.update assert_approx_equals: Red channel of the pixel at (50, 25) expected 0 +/- 2 but got 255
